ml4db/requirements.txt
absl-py
adabound
alembic
antlr4-python3-runtime
appnope
argon2-cffi
argon2-cffi-bindings
asttokens
attrs
autopage
backcall
beautifulsoup4
black
bleach
cachetools
certifi
cffi
charset-normalizer
click
cliff
cmaes
cmd2
colorlog
cycler
debugpy
decorator
defusedxml
entrypoints
executing
faiss-cpu
fastjsonschema
flake8
fonttools
google-auth
google-auth-oauthlib
grpcio
hydra-core
idna
importlib-metadata
ipykernel
ipython
ipython-genutils
ipywidgets
jedi
Jinja2
joblib
jsonschema
jupyter
jupyter-console
jupyter-core
jupyter_client
jupyterlab-pygments
jupyterlab-widgets
kiwisolver
llvmlite
lxml
Mako
Markdown
MarkupSafe
matplotlib
matplotlib-inline
mccabe
mistune
mypy-extensions
nbclient
nbconvert
nbformat
nest-asyncio
notebook
numba
numpy
oauthlib
omegaconf
optuna
packaging
pandas
pandocfilters
parso
pathspec
pbr
pexpect
pickleshare
Pillow
platformdirs
prettytable
prometheus-client
prompt-toolkit
protobuf
psutil
ptyprocess
pure-eval
pyasn1
pyasn1-modules
pycodestyle
pycparser
pyflakes
Pygments
pynndescent
pyparsing
pyperclip
pyrsistent
pytorch_metric_learning
python-dateutil
pytz
PyYAML
pyzmq
qtconsole
QtPy
record-keeper
requests
requests-oauthlib
rsa
scikit-learn
scipy
seaborn
Send2Trash
six
soupsieve
SQLAlchemy
stack-data
stevedore
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
terminado
threadpoolctl
tinycss2
tomli
torch
torch-tb-profiler
torchvision
tornado
tqdm
traitlets
typing_extensions
umap-learn
urllib3
wcwidth
webencodings
Werkzeug
widgetsnbextension
zipp
